requirements.txt
chardet==2.3.0
cookiecutter==1.4.0
Django==1.10.2
djangorestframework==3.4.7
future==0.15.2
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
jinja2-time==0.2.0
MarkupSafe==0.23
nba-py==0.1a1
numpy==1.11.2
pandas==0.19.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
poyo==0.4.0
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.7
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
whichcraft==0.4.0
django_extensions==1.7.4
model-mommy==1.3.0
